i
ip

+++ point.java
@@ -23,7 +23,7 @@ public class point{
     public boolean isCollinear(point p1, point p2,point p3){
         double slope;
         if(p2.getX()-p1.getX()!=0){
-            slope = (p2.getY()-p1.getY())/(p2.getX()-p1.getX());
+            slope = (double)(p2.getY()-p1.getY())/(p2.getX()-p1.getX());
         }else{
             if(p2.getX()==p3.getX()){
                 return true;
@@ -32,7 +32,8 @@ public class point{
             }
         }
         //y-y1 = m(x-x1)
-        double yintercept = slope*(0-p1.getX())+p1.getY();
+        double yintercept = (double)slope*(0-p1.getX())+p1.getY();
+    
         if((yintercept+slope*p3.getX()<p3.getY()+0.01)&&(yintercept+slope*p3.getX()>p3.getY()-0.01)){
             return true;
         }else{
